% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep_function.R
\name{grep_count}
\alias{grep_count}
\title{Count lines matching pattern}
\usage{
grep_count(pattern, file)
}
\arguments{
\item{pattern}{Regex pattern}

\item{file}{File path}
}
\value{
Line count
}
\description{
Count lines matching pattern
}
