% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep_function.R
\name{grep_read}
\alias{grep_read}
\title{Read lines matching pattern}
\usage{
grep_read(pattern, file)
}
\arguments{
\item{pattern}{Regex pattern}

\item{file}{File path}
}
\value{
data.table
}
\description{
Read lines matching pattern
}
